ヘッダーから登録機能
