+++ app/views/cards/new.html.haml
@@ -1,4 +1,4 @@
-= form_tag(pay_card_index_path, method: :post, id: 'charge-form',  name: "inputForm") do
+= form_tag(cards_path, method: :post, id: 'charge-form',  name: "inputForm") do
   %label カード番号
   = text_field_tag "number", "", class: "number", placeholder: "半角数字のみ" ,maxlength: "16", type: "text", id: "card_number"
   %br
@@ -20,8 +20,6 @@
   %span 月/
   %select#exp_year{name: "exp_year", type: "text"}
     %option{value: ""} --
-    %option{value: "2019"}19
-    %option{value: "2020"}20
     %option{value: "2021"}21
     %option{value: "2022"}22
     %option{value: "2023"}23
