+++ app/views/cards/show.html.haml
@@ -1,11 +1,11 @@
 %label 登録クレジットカード情報
-%br
+-# %br
 -# = "**** **** **** " + @default_card_information.last4
-%br
+-# %br
 -# - exp_month = @default_card_information.exp_month.to_s
 -# - exp_year = @default_card_information.exp_year.to_s.slice(2,3)
 -# = exp_month + " / " + exp_year
-= form_tag(delete_card_index_path, method: :post, id: 'charge-form',  name: "inputForm") do
-  %input{ type: "hidden", name: "card_id", value: "" }
-  %button 削除する
+-# = form_tag(card_path, method: :DELETE, id: 'charge-form',  name: "inputForm") do
+-#   %input{ type: "hidden", name: "card_id", value: "" }
+-#   %button 削除する
 = link_to "戻る","/"
