+++ config/routes.rb
@@ -7,11 +7,5 @@ Rails.application.routes.draw do
   resources :users, only: :show
   resources :items
 
-  resources :card, only: [:new, :show] do
-    collection do
-      post 'show', to: 'card#show'
-      post 'pay', to: 'card#pay'
-      post 'delete', to: 'card#delete'
-    end
-  end
+  resources :cards
 end
